python-backend/requirements.txt
openai>=1.0.0
pydantic
fastapi
uvicorn
python-dotenv
